Jeg har rykket metoderne til TimeCalculator, gjort den til en Spring Component og dependency injected den der, hvor den skal bruges
